fix: correct input parameters for https.request()
Prior to this change I got an error message like

```
(node:17319) UnhandledPromiseRejectionWarning: TypeError [ERR_INVALID_ARG_TYPE]: The "listener" argument must be of type Function. Received type object
    at ClientRequest.once (events.js:288:11)
    at new ClientRequest (_http_client.js:141:10)
    at Object.request (https.js:272:10)
    at resolve (/home/vmx/src/pl/package-table/src/badges.js:4:21)
    at new Promise (<anonymous>)
    at badgeExists (/home/vmx/src/pl/package-table/src/badges.js:3:28)
    at Object.ciJenkins (/home/vmx/src/pl/package-table/src/badges.js:44:13)
    at data.columns.map.col (/home/vmx/src/pl/package-table/src/index.js:18:62)
    at Array.map (<anonymous>)
    at generatePackageRow (/home/vmx/src/pl/package-table/src/index.js:18:33)
(node:17319) UnhandledPromiseRejectionWarning: Unhandled promise rejection. This error originated either by throwing inside of an async function without a catch block, or by rejecting a promise which was not handled with .catch(). (rejection id: 13)
(node:17319) [DEP0018] DeprecationWarning: Unhandled promise rejections are deprecated. In the future, promise rejections that are not handled will terminate the Node.js process with a non-zero exit code.
```

The problem was that the input parameters for `https.request()` were wrong.
It takes an URL or options, not both.